Moved all the models into a new folder also named models
