+++ ProjectJoao/app/src/main/java/com/example/android/projectjoao/model/ListResponse.java
@@ -1,6 +1,7 @@
-package com.example.android.projectjoao;
+package com.example.android.projectjoao.model;
 
 import java.util.List;
+
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
