+++ ProjectJoao/app/src/main/java/com/example/android/projectjoao/model/ShowResponse.java
@@ -1,0 +1,20 @@
+package com.example.android.projectjoao.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class ShowResponse {
+
+    @SerializedName("data")
+    @Expose
+    private User data;
+
+    public User getData() {
+        return data;
+    }
+
+    public void setData(User data) {
+        this.data = data;
+    }
+
+}
